clean + readme + examples
