+++ examples/Callable_from_qua/calibration_db.json
@@ -10,49 +10,49 @@
             "mode_id": 1,
             "lo_freq": 4500000000.0,
             "gain": 0.0,
-            "latest": 17
+            "latest": 25
         },
         "2": {
             "mode_id": 1,
             "lo_freq": 4750000000.0,
             "gain": 0.0,
-            "latest": 18
+            "latest": 26
         },
         "3": {
             "mode_id": 1,
             "lo_freq": 5000000000.0,
             "gain": 0.0,
-            "latest": 19
+            "latest": 27
         },
         "4": {
             "mode_id": 1,
             "lo_freq": 5250000000.0,
             "gain": 0.0,
-            "latest": 20
+            "latest": 28
         },
         "5": {
             "mode_id": 1,
             "lo_freq": 5500000000.0,
             "gain": 0.0,
-            "latest": 21
+            "latest": 29
         },
         "6": {
             "mode_id": 1,
             "lo_freq": 5750000000.0,
             "gain": 0.0,
-            "latest": 22
+            "latest": 30
         },
         "7": {
             "mode_id": 1,
             "lo_freq": 6000000000.0,
             "gain": 0.0,
-            "latest": 23
+            "latest": 31
         },
         "8": {
             "mode_id": 1,
             "lo_freq": 6250000000.0,
             "gain": 0.0,
-            "latest": 24
+            "latest": 32
         }
     },
     "lo_cal": {
@@ -295,6 +295,86 @@
             "temperature": 60.0625,
             "timestamp": 1700762873.7035182,
             "method": "auto"
+        },
+        "25": {
+            "i0": -0.003829600222011526,
+            "q0": 0.03680077145715791,
+            "lo_mode_id": 1,
+            "dc_gain": -0.013096190802231233,
+            "dc_phase": 0.13487299201801253,
+            "temperature": 60.0,
+            "timestamp": 1700840218.7019324,
+            "method": "auto"
+        },
+        "26": {
+            "i0": 0.005747511398680342,
+            "q0": 0.02756409567906909,
+            "lo_mode_id": 2,
+            "dc_gain": -0.00012915641031807645,
+            "dc_phase": 0.14520768413579307,
+            "temperature": 60.0,
+            "timestamp": 1700840238.5398724,
+            "method": "auto"
+        },
+        "27": {
+            "i0": 0.0004904712174455584,
+            "q0": 0.02692410739287672,
+            "lo_mode_id": 3,
+            "dc_gain": 0.002761409377456392,
+            "dc_phase": 0.16112626682842576,
+            "temperature": 60.0,
+            "timestamp": 1700840258.1165664,
+            "method": "auto"
+        },
+        "28": {
+            "i0": 0.00415677548700119,
+            "q0": 0.019847656527453576,
+            "lo_mode_id": 4,
+            "dc_gain": 0.009447273395866373,
+            "dc_phase": 0.13035163294469226,
+            "temperature": 60.0,
+            "timestamp": 1700840277.8931394,
+            "method": "auto"
+        },
+        "29": {
+            "i0": -0.0014631578366805632,
+            "q0": 0.016815630210953985,
+            "lo_mode_id": 5,
+            "dc_gain": -0.0026216714555525877,
+            "dc_phase": 0.028930970993157876,
+            "temperature": 60.0625,
+            "timestamp": 1700840298.014474,
+            "method": "auto"
+        },
+        "30": {
+            "i0": -0.0006095370812436661,
+            "q0": 0.013420599260907325,
+            "lo_mode_id": 6,
+            "dc_gain": -0.005202656285122509,
+            "dc_phase": -0.013497250775683277,
+            "temperature": 60.0625,
+            "timestamp": 1700840317.5702689,
+            "method": "auto"
+        },
+        "31": {
+            "i0": -0.003039758063583113,
+            "q0": 0.01732737214170537,
+            "lo_mode_id": 7,
+            "dc_gain": -0.0019316200179706478,
+            "dc_phase": -0.02298930909901791,
+            "temperature": 60.0625,
+            "timestamp": 1700840337.0353808,
+            "method": "auto"
+        },
+        "32": {
+            "i0": 0.0026409968942230633,
+            "q0": 0.0085930993831169,
+            "lo_mode_id": 8,
+            "dc_gain": -0.004633310764357912,
+            "dc_phase": -0.09078345080152898,
+            "temperature": 60.0625,
+            "timestamp": 1700840356.2732127,
+            "method": "auto"
         }
     },
     "if_modes": {
@@ -341,42 +421,42 @@
         "9": {
             "lo_mode_id": 1,
             "if_freq": 126000000.0,
-            "latest": 17
+            "latest": 25
         },
         "10": {
             "lo_mode_id": 2,
             "if_freq": 126000000.0,
-            "latest": 18
+            "latest": 26
         },
         "11": {
             "lo_mode_id": 3,
             "if_freq": 126000000.0,
-            "latest": 19
+            "latest": 27
         },
         "12": {
             "lo_mode_id": 4,
             "if_freq": 126000000.0,
-            "latest": 20
+            "latest": 28
         },
         "13": {
             "lo_mode_id": 5,
             "if_freq": 126000000.0,
-            "latest": 21
+            "latest": 29
         },
         "14": {
             "lo_mode_id": 6,
             "if_freq": 126000000.0,
-            "latest": 22
+            "latest": 30
         },
         "15": {
             "lo_mode_id": 7,
             "if_freq": 126000000.0,
-            "latest": 23
+            "latest": 31
         },
         "16": {
             "lo_mode_id": 8,
             "if_freq": 126000000.0,
-            "latest": 24
+            "latest": 32
         }
     },
     "if_cal": {
@@ -570,6 +650,70 @@
             "phase": -0.11139509172046443,
             "temperature": 60.0625,
             "timestamp": 1700762873.716873,
+            "method": "auto"
+        },
+        "25": {
+            "if_mode_id": 9,
+            "gain": -0.018392773588505343,
+            "phase": 0.1619538630743425,
+            "temperature": 60.0,
+            "timestamp": 1700840218.713932,
+            "method": "auto"
+        },
+        "26": {
+            "if_mode_id": 10,
+            "gain": 0.0008211270008152138,
+            "phase": 0.16919208862639815,
+            "temperature": 60.0,
+            "timestamp": 1700840238.5516362,
+            "method": "auto"
+        },
+        "27": {
+            "if_mode_id": 11,
+            "gain": 0.0029181038127522813,
+            "phase": 0.1792769381539723,
+            "temperature": 60.0,
+            "timestamp": 1700840258.1310575,
+            "method": "auto"
+        },
+        "28": {
+            "if_mode_id": 12,
+            "gain": 0.0056985078736210264,
+            "phase": 0.1444247784447213,
+            "temperature": 60.0,
+            "timestamp": 1700840277.9022455,
+            "method": "auto"
+        },
+        "29": {
+            "if_mode_id": 13,
+            "gain": -0.004928794350060633,
+            "phase": 0.03905511239013761,
+            "temperature": 60.0625,
+            "timestamp": 1700840298.0256302,
+            "method": "auto"
+        },
+        "30": {
+            "if_mode_id": 14,
+            "gain": -0.005139359186123539,
+            "phase": -0.004070144763211948,
+            "temperature": 60.0625,
+            "timestamp": 1700840317.5802753,
+            "method": "auto"
+        },
+        "31": {
+            "if_mode_id": 15,
+            "gain": -0.004128737728761285,
+            "phase": -0.014932905625431711,
+            "temperature": 60.0625,
+            "timestamp": 1700840337.0482764,
+            "method": "auto"
+        },
+        "32": {
+            "if_mode_id": 16,
+            "gain": -0.008999686674574691,
+            "phase": -0.11137481028898222,
+            "temperature": 60.0625,
+            "timestamp": 1700840356.285234,
             "method": "auto"
         }
     }
