+++ examples/Callable_from_qua/configuration_octave.py
