+++ examples/Callable_from_qua/set_octave.py
