+++ examples/Callable_from_qua/test_live_plot.py
@@ -1,11 +1,12 @@
 from qm.qua import *
 from qm.QuantumMachinesManager import QuantumMachinesManager
-from qm import SimulationConfig
 from configuration import *
 import matplotlib.pyplot as plt
 import warnings
 from qm import generate_qua_script
-from callable_from_qua2 import CallableFromQUA, run_local
+
+from callable_from_qua import program, run_local
+
 
 warnings.filterwarnings("ignore")
 
@@ -25,19 +26,16 @@ def set_lo_power(q: str, qm, value):
 qmm = QuantumMachinesManager(host="172.16.33.101", cluster_name="Cluster_83")
 # Open a quantum machine
 qm = qmm.open_qm(config)
-# Declare the Callable from QUA framework
-batches = CallableFromQUA()
 # Define your QUA program with the declare_all() and do_stream_processing() lines
 with program() as prog:
     n = declare(int)
     n2 = declare(int)
     I = declare(fixed)
     I_st = declare_stream()
-    batches.declare_all()  # TODO: need batches declaration
     with for_(n2, 0, n2 < 2, n2 + 1):
-        set_lo_power(batches=batches, q="QUBIT", qm=qm, value=n2) # TODO: remove batches from here?
+        set_lo_power(q="QUBIT", qm=qm, value=n2)
         with for_(n, 0, n < 3, n + 1):
-            set_lo_freq(batches=batches, q="AOM", qm=qm, value=n)
+            set_lo_freq(q="AOM", qm=qm, value=n)
             play("cw", "AOM")
             measure(
                 "readout",
@@ -49,8 +47,18 @@ with program() as prog:
 
     with stream_processing():
         I_st.save_all("I")
-    batches.do_stream_processing()  # TODO: need batches stream processing
+        
 # Execute the QUA program using the callable from QUA framework
-job = batches.execute(qm, prog)
+fig = plt.figure()
+def live_plot(res_handles):
+    data = res_handles.get("I").fetch_all()["value"]
+    plt.cla()
+    plt.plot(data, "ko")
+    plt.pause(0.01)
+
+with prog.local_run(qm, funcs=[live_plot]):
+    job = qm.execute(prog)
+
+# job = batches.execute(qm, prog)  # TODO Make this unnecessary
 
 print(generate_qua_script(prog, config)[:1000])
