--- qualang_tools/addons/callable_from_qua/my_run_local.py
@@ -1,9 +1,0 @@
-from my_BatchJob import BatchJob
-
-
-# def run_local(func):
-#     def wrapper(batches: BatchJob, *args, **kwargs):
-#         batches.local_run(func, *args, **kwargs)
-#     return wrapper
-
-
