--- qualang_tools/addons/callable_from_qua/run_local.py
@@ -1,7 +1,0 @@
-from my_BatchJob import BatchJob
-
-
-def run_local(func):
-    def wrapper(batches: BatchJob, *args, **kwargs):
-        batches.local_run(func, *args, **kwargs)
-    return wrapper
